[Infra/Feat] 무중단 Blue/Green 배포 스크립트 작성 S12P31D103-133

See merge request s12-final/S12P31D103!167

+++ Android/app/src/main/java/com/hogumiwarts/lumos/ui/common/DeviceGridHomeSection.kt
@@ -36,7 +36,7 @@ import com.hogumiwarts.lumos.ui.theme.nanum_square_neo
 @Composable
 fun DeviceGridHomeSection(
     devices: List<MyDevice>,
-    selectedDeviceId: Int? = null,
+    selectedDeviceId: Long? = null,
     onDeviceClick: (MyDevice) -> Unit = {},
     onToggleDevice: (MyDevice) -> Unit = {}
 ) {
@@ -77,7 +77,9 @@ fun DeviceGridHomeSection(
                     DeviceRoutineCard(
                         modifier = Modifier
                             .fillMaxSize()
-                            .clickable { onDeviceClick(device) },
+                            .clickable {
+                                onDeviceClick(device)
+                                       },
                         showToggle = true,
                         cardTitle = device.deviceName,
                         cardSubtitle = device.deviceType.categoryName,
