+++ Android/data/src/main/java/com/hogumiwarts/data/entity/remote/Response/routine/RoutineCreateResponse.kt
@@ -1,9 +1,9 @@
 package com.hogumiwarts.data.entity.remote.Response.routine
 
 data class RoutineCreateResponse(
-    val routineId: Int,
+    val routineId: Long,
     val memberId: Int,
-    val gestureId: Int,
+    val gestureId: Long,
     val routineName: String,
     val routineIcon: String,
     val devices: List<RoutineDeviceData>
