+++ Android/data/src/main/java/com/hogumiwarts/data/entity/remote/Response/routine/RoutineResponse.kt
@@ -1,13 +1,13 @@
 package com.hogumiwarts.data.entity.remote.Response.routine
 
 data class RoutineResponse(
-    val status: Int,
+    val status: Long,
     val message: String,
     val data: List<RoutineData>
 )
 
 data class RoutineData(
-    val routineId: Int,
+    val routineId: Long,
     val routineName: String,
     val routineIcon: String,
     val gestureName: String
